Enhance Makefile and improve error handling in API
- Added GIT_SHA variable to Makefile for tagging Docker images with the current commit SHA.
- Updated Docker build commands to tag migration images with both 'latest' and the commit SHA.
- Improved error handling in the Gemini client by returning the original error when failing to open a file.
- Enhanced video analysis task logging in the worker to include retry counts and handle maximum retries.
- Added IAM permissions for the Cloud Run service account in a new Terraform configuration file.

+++ api/internal/worker/handler.go
@@ -90,7 +90,20 @@ func (w *Worker) HandleVideoAnalysisTask(ctx context.Context, t *asynq.Task) err
 		return fmt.Errorf("json.Unmarshal failed: %v: %w", err, asynq.SkipRetry)
 	}
 
-	logger.Log.Info("Processing video analysis", zap.String("session_id", p.SessionID), zap.String("file_path", p.FilePath))
+	retryCount, ok := asynq.GetRetryCount(ctx)
+	if !ok {
+		retryCount = 0
+	}
+
+	logger.Log.Info("Processing video analysis",
+		zap.String("session_id", p.SessionID),
+		zap.String("file_path", p.FilePath),
+		zap.Int("retry_count", int(retryCount)))
+
+	if retryCount >= 3 {
+		logger.Log.Error("Max retries reached. Skipping analysis.")
+		return asynq.SkipRetry
+	}
 
 	// Determine file path (download from GCS if needed)
 	localFilePath := p.FilePath
@@ -115,6 +128,12 @@ func (w *Worker) HandleVideoAnalysisTask(ctx context.Context, t *asynq.Task) err
 
 	analysis, geminiFile, err := geminiClient.AnalyzeVideo(ctx, localFilePath, AnalysisPrompt)
 
+	// retry if analysis is empty
+	if analysis == "" {
+		logger.Log.Warn("Analysis failed. Retrying...", zap.Error(err))
+		return fmt.Errorf("analysis is empty")
+	}
+
 	// Clean up local file
 	defer func() {
 		if err := os.Remove(localFilePath); err != nil {
@@ -132,6 +151,11 @@ func (w *Worker) HandleVideoAnalysisTask(ctx context.Context, t *asynq.Task) err
 	}
 
 	if err != nil {
+		if os.IsNotExist(err) {
+			logger.Log.Error("File not found. Skipping analysis.", zap.Error(err))
+			return asynq.SkipRetry
+		}
+
 		logger.Log.Error("Analysis failed", zap.Error(err))
 		// Save failure to DB
 		w.DB.Create(&db.AnalysisResult{
